+++ ios/Classes/TiAdmobView.h
@@ -8,7 +8,6 @@
 #import "TiUIView.h"
 #import "GADBannerView.h"
 #import "GADBannerViewDelegate.h"
-#import <AdSupport/ASIdentifierManager.h>
 
 @interface TiAdmobView : TiUIView<GADBannerViewDelegate> {