+++ ios/documentation/changelog.md
@@ -1,5 +1,10 @@
 # Change Log
 <pre>
+v1.6.0  [TIMODOPEN-427] Updating Admob SDK to 6.10.0, Building with TiSDK 3.2.3.GA, update documentation
+        Replaced `publisherId` with `adUnitId`
+        Deprecated `testing`
+        Added `testDevices`
+
 v1.5.0  [TIMODOPEN-242] Updating Admob SDK to 6.4.2 (Removed all uses of UDID)
 
 v1.4.0  [TIMODOPEN-212] Updating Admob SDK to 6.2.1, Building with TiSDK 2.1.3.GA, update documentation